pyproject.toml
[tool.poetry]
name = "expden"
version = "0.1.0"
description = "Library to compute the experienced density"
authors = ["Nico Forteza  <nicoforteza@gmail.com>"]
maintainers = ["Nico Forteza  <nicoforteza@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.16"
numba = "0.56.4"
geopandas = "0.12.1"
xarray-spatial = "0.3.5"
rasterstats = "0.17.0"
rioxarray = "0.13.1"
geocube = "0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

